- Camelized property-Names, id->Long

+++ datamodel/src/main/java/com/apt/modernization/datamodel/repository/QuotesRepository.java
@@ -3,6 +3,6 @@ package com.apt.modernization.datamodel.repository;
 import com.apt.modernization.datamodel.document.Quotes;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface QuotesRepository extends MongoRepository<Quotes, Integer> {
-	Quotes findById(int id);
+public interface QuotesRepository extends MongoRepository<Quotes, Long> {
+	Quotes findById(long id);
 }
